+++ core/components/crosslinks/docs/changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2019-01-04
+### Added
+- Don't replace crosslinks in definable html tags
+- Don't replace crosslinks in all html tag attributes
+
 ## [1.1.0] - 2018-12-02
 ### Added
 - Don't replace crosslinks in definable html attributes